Example of using coparison logic
The movie theater has cinema halls that can accommodate a certain number of viewers each. Figure out if a movie theater can hold a given number of viewers that plan to visit it on a particular day.

The input format

The first line is number of halls, the second line is their capacity, and the third line is the planned number of viewers.

The output format

True or False.